ARCH-3: Moved models back to Models dir
