+++ WineUp/WineUp/Views/CatalogFiltersBar/CatalogFiltersBarView.swift
@@ -30,16 +30,6 @@ struct CatalogFiltersBarView: View {
     }
 }
 
-// MARK: - Model
-
-extension CatalogFiltersBarView {
-    struct Item: Identifiable {
-        var id = UUID()
-        /// Filter name
-        var title: String
-    }
-}
-
 // MARK: - Displaying Items
 
 private extension CatalogFiltersBarView {
