+++ WineUp/WineUp/Views/Filters/PriceFilter/PriceFilterView.swift
@@ -68,17 +68,6 @@ struct PriceFilterView: View {
     }
 }
 
-// MARK: - Model
-
-extension PriceFilterView {
-    /// Predefined price interval like `500-1000`
-    struct Item: Identifiable {
-        var id = UUID()
-        /// Text representation like `500-1000`
-        var title: String
-    }
-}
-
 // MARK: - Helpers
 
 private extension PriceFilterView {
